Don't run releases.py (just make releases.sh responsible for upload) Only delete the file from the bucket if it's not being updated

+++ jira/releases.sh
@@ -10,26 +10,29 @@ if [ "" == "${S3_BUCKET}" ]; then
 	exit 1
 fi
 
-python releases.py
-
 aws s3 --profile ${AWS_PROFILE} ls s3://${S3_BUCKET}/releases/ | awk '{ print $4 "\t" $3 }' | sort > 1.txt
 
 cd releases.production
 
 for file in *.json; do
-  gzip -c ${file} > ${file}.gz
+	gzip -c ${file} > ${file}.gz
 done
 
 cd -
 
 ls -l releases.production/*.gz | sed 's/\.gz$//g' | sed 's@releases.production/@@g' | awk '{ print $9 "\t" $5 }' | sort > 2.txt
 
-for file in $(diff 1.txt 2.txt | grep '<' | awk '{ print $2 }'); do
-  aws s3 --profile ${AWS_PROFILE} rm s3://${S3_BUCKET}/releases/${file}
+diff 1.txt 2.txt | grep '<' | awk '{ print $2 }' > 3.txt
+diff 1.txt 2.txt | grep '>' | awk '{ print $2 }' > 4.txt
+
+for file in $(cat 3.txt); do
+	if [ "" == "$(grep "^${file}$" 4.txt)" ]; then
+		aws s3 --profile ${AWS_PROFILE} rm s3://${S3_BUCKET}/releases/${file}
+	fi
 done
 
-for file in $(diff 1.txt 2.txt | grep '>' | awk '{ print $2 }'); do
-  aws s3 --profile ${AWS_PROFILE} cp releases.production/${file}.gz "s3://${S3_BUCKET}/releases/${file}" --acl public-read --metadata-directive REPLACE --content-encoding gzip
+for file in $(cat 4.txt); do
+	aws s3 --profile ${AWS_PROFILE} cp releases.production/${file}.gz "s3://${S3_BUCKET}/releases/${file}" --acl public-read --metadata-directive REPLACE --content-encoding gzip
 done
 
 gzip -c releases.production.json > releases.json.gz
@@ -42,4 +45,4 @@ for file in releases.html releases.js; do
   rm ${file}.gz
 done
 
-rm 1.txt 2.txt releases.production/*.gz
+rm 1.txt 2.txt 3.txt 4.txt releases.production/*.gz
